raspberry_pi/requirements_pi.txt
# Core ML/AI dependencies cho Raspberry Pi
ultralytics>=8.0.0
torch>=2.0.0

# Computer Vision
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# NCNN for Raspberry Pi optimization
ncnn>=1.0.0

# UART communication
pyserial>=3.5

# MQTT communication
paho-mqtt>=1.6.0

# Image processing
Pillow>=10.0.0

# Performance monitoring
psutil>=5.9.0

# Threading và multiprocessing
threading
multiprocessing 